requirements.txt
langchain==0.2.7
langchain_community==0.2.7
langchain_core==0.2.18
langchain_groq==0.1.6
langchain_huggingface==0.0.3
python-dotenv==1.0.1
pypdf==4.2.0
faiss-cpu==1.8.0.post1
